Initial release 4

+++ kernel/include/platform/link.h
@@ -11,18 +11,20 @@ Author: myaut
 #ifndef LINK_H_
 #define LINK_H_
 
+#include <types.h>
+
 /*
  * Linker address
  * Defines each kernel
  */
 #define kernel_text_start	   0x0000000
-extern unsigned long kernel_text_end;
-extern unsigned long kernel_data_start;
-extern unsigned long kernel_data_end;
-extern unsigned long kernel_bss_start;
-extern unsigned long kernel_bss_end;
-extern unsigned long kernel_kip_start;
-extern unsigned long kernel_kip_end;
+extern uint32_t kernel_text_end;
+extern uint32_t kernel_data_start;
+extern uint32_t kernel_data_end;
+extern uint32_t kernel_bss_start;
+extern uint32_t kernel_bss_end;
+extern uint32_t kernel_kip_start;
+extern uint32_t kernel_kip_end;
 extern void kernel_stack_addr(void);
 
 #define __BSS 			__attribute__ ((section(".bss")))
